requirements.txt
requests
pandas
python-dotenv
huggingface_hub